Initial DX11 setup, took more than 2hours, hence no commit yesterday.
Added a renderstate to alternate between vsync and fullscreen from the app side.

Added a template for memorystack allocation (will have to check the compilation performance on this later)

Change location of gamestate into PlatformLayer. The idea is that this should act like a 'hub' for the game and platformlayer to delegate memory and data to the different sub systems within the app.

Signed-off-by: patrik <patrik.fjellstedt@gmail.com>